tried to move planet

+++ Assets/Scripts/GravitySimulation/KeplerParams.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class KeplerParams
 {
-    public float gravityParam;
-    public float semiMajorAxis;
-    public float eccentricity;
-    public float argumentOfPeriapsis;
-    public float longitudeOfAscendingNode;
-    public float inclination;
-    public float trueAnomalyAtEpoch;
-    public float timeToPeriapsis;
-    public float AP;
+    public double gravityParam;
+    public double semiMajorAxis;
+    public double eccentricity;
+    public double argumentOfPeriapsis;
+    public double longitudeOfAscendingNode;
+    public double inclination;
+    public double trueAnomalyAtEpoch;
+    public double timeToPeriapsis;
+    public double AP;
 }
